Rename ColorSchemeManager to ColorSchemeRegistry
