+++ superset/assets/src/modules/colors/CategoricalColorNamespace.js
@@ -1,5 +1,5 @@
 import CategoricalColorScale from './CategoricalColorScale';
-import getCategoricalSchemeManager from './CategoricalSchemeManagerSingleton';
+import getCategoricalSchemeRegistry from './CategoricalSchemeRegistrySingleton';
 
 class CategoricalColorNamespace {
   constructor(name) {
@@ -9,13 +9,13 @@ class CategoricalColorNamespace {
   }
 
   getScale(schemeName) {
-    const name = schemeName || getCategoricalSchemeManager().getDefaultSchemeName();
+    const name = schemeName || getCategoricalSchemeRegistry().getDefaultSchemeName();
     const scale = this.scales[name];
     if (scale) {
       return scale;
     }
     const newScale = new CategoricalColorScale(
-      getCategoricalSchemeManager().getScheme(name).colors,
+      getCategoricalSchemeRegistry().getScheme(name).colors,
       this.forcedItems,
     );
     this.scales[name] = newScale;
