--- superset/assets/src/modules/colors/CategoricalSchemeManagerSingleton.js
@@ -1,8 +1,0 @@
-import ColorSchemeManager from './ColorSchemeManager';
-import makeSingleton from '../../utils/makeSingleton';
-
-class CategoricalSchemeManager extends ColorSchemeManager {}
-
-const getInstance = makeSingleton(CategoricalSchemeManager);
-
-export default getInstance;
