+++ superset/assets/src/modules/colors/CategoricalSchemeRegistrySingleton.js
@@ -1,0 +1,8 @@
+import ColorSchemeRegistry from './ColorSchemeRegistry';
+import makeSingleton from '../../utils/makeSingleton';
+
+class CategoricalSchemeRegistry extends ColorSchemeRegistry {}
+
+const getInstance = makeSingleton(CategoricalSchemeRegistry);
+
+export default getInstance;
