--- superset/assets/src/modules/colors/SequentialSchemeManagerSingleton.js
@@ -1,8 +1,0 @@
-import ColorSchemeManager from './ColorSchemeManager';
-import makeSingleton from '../../utils/makeSingleton';
-
-class SequentialSchemeManager extends ColorSchemeManager {}
-
-const getInstance = makeSingleton(SequentialSchemeManager);
-
-export default getInstance;
