+++ superset/assets/src/modules/colors/SequentialSchemeRegistrySingleton.js
@@ -1,0 +1,8 @@
+import ColorSchemeRegistry from './ColorSchemeRegistry';
+import makeSingleton from '../../utils/makeSingleton';
+
+class SequentialSchemeRegistry extends ColorSchemeRegistry {}
+
+const getInstance = makeSingleton(SequentialSchemeRegistry);
+
+export default getInstance;
